<refact> :타입 정리 및 에러 핸들링 추가
