+++ src/components/auth/CheckTerm/CheckTerm.tsx
