+++ src/components/auth/CheckTerm/style.ts
